OEL-740: Refactor. Move classes to own folder.
