+++ tests/src/Kernel/AbstractKernelTestBase.php
@@ -36,7 +36,7 @@ abstract class AbstractKernelTestBase extends KernelTestBase {
   /**
    * {@inheritdoc}
    */
-  protected function setUp() {
+  protected function setUp(): void {
     parent::setUp();
 
     $this->installEntitySchema('user');
